+++ finetrainers/trainer/base.py
@@ -1,10 +1,13 @@
 import contextlib
 import functools
+import os
 from typing import Callable, List, Tuple
 
 import torch
+import torch.backends
 from diffusers.hooks import HookRegistry, ModelHook
 
+from finetrainers import parallel, logging
 from finetrainers.args import BaseArgsType
 from finetrainers.logging import get_logger
 from finetrainers.models.attention_dispatch import AttentionProvider, _AttentionProviderRegistry
@@ -21,6 +24,9 @@ class Trainer:
 
         self._module_name_providers_training = _parse_attention_providers(args.attn_provider_training)
         self._module_name_providers_inference = _parse_attention_providers(args.attn_provider_inference)
+
+        self._init_distributed()
+        self._init_config_options()
 
     @contextlib.contextmanager
     def attention_provider_ctx(self, training: bool = True):
@@ -66,6 +72,47 @@ class Trainer:
         for module in module_providers_dict.keys():
             registry: HookRegistry = module._diffusers_hook
             registry.remove_hook(_LATEST_ACTIVE_MODULE_HOOK)
+    
+    def _init_distributed(self) -> None:
+        world_size = int(os.environ.get("WORLD_SIZE", torch.cuda.device_count()))
+
+        # TODO(aryan): handle other backends
+        backend_cls: parallel.ParallelBackendType = parallel.get_parallel_backend_cls(self.args.parallel_backend)
+        self.state.parallel_backend = backend_cls(
+            world_size=world_size,
+            pp_degree=self.args.pp_degree,
+            dp_degree=self.args.dp_degree,
+            dp_shards=self.args.dp_shards,
+            cp_degree=self.args.cp_degree,
+            tp_degree=self.args.tp_degree,
+            backend="nccl",
+            timeout=self.args.init_timeout,
+            logging_dir=self.args.logging_dir,
+            output_dir=self.args.output_dir,
+            gradient_accumulation_steps=self.args.gradient_accumulation_steps,
+        )
+
+        if self.args.seed is not None:
+            self.state.parallel_backend.enable_determinism(self.args.seed)
+    
+    def _init_logging(self) -> None:
+        logging._set_parallel_backend(self.state.parallel_backend)
+        logging.set_dependency_log_level(self.args.verbose, self.state.parallel_backend.is_local_main_process)
+        logger.info("Initialized FineTrainers")
+
+    def _init_trackers(self) -> None:
+        # TODO(aryan): handle multiple trackers
+        trackers = [self.args.report_to]
+        experiment_name = self.args.tracker_name or "finetrainers-experiment"
+        self.state.parallel_backend.initialize_trackers(
+            trackers, experiment_name=experiment_name, config=self._get_training_info(), log_dir=self.args.logging_dir
+        )
+    
+    def _init_config_options(self) -> None:
+        # Enable TF32 for faster training on Ampere GPUs: https://pytorch.org/docs/stable/notes/cuda.html#tensorfloat-32-tf32-on-ampere-devices
+        if self.args.allow_tf32 and torch.cuda.is_available():
+            torch.backends.cuda.matmul.allow_tf32 = True
+        torch.set_float32_matmul_precision(self.args.float32_matmul_precision)
 
 
 class LatestActiveModuleHook(ModelHook):
