+++ finetrainers/trainer/sft_trainer/trainer.py
@@ -8,7 +8,6 @@ from typing import Any, Dict, Iterable, List, Optional, Union
 
 import datasets.distributed
 import torch
-import torch.backends
 import wandb
 from diffusers import DiffusionPipeline
 from diffusers.hooks import apply_layerwise_casting
@@ -69,9 +68,6 @@ class SFTTrainer(Trainer):
 
         # Checkpoint manager
         self.checkpointer = None
-
-        self._init_distributed()
-        self._init_config_options()
 
         # Perform any patches that might be necessary for training to work as expected
         patches.perform_patches_for_training(self.args, self.state.parallel_backend)
@@ -152,15 +148,13 @@ class SFTTrainer(Trainer):
         model_specification = self.model_specification
 
         if parallel_backend.context_parallel_enabled:
-            raise NotImplementedError(
-                "Context parallelism is not supported yet. This will be supported in the future."
-            )
+            parallel_backend.apply_context_parallel(self.transformer, parallel_backend.get_mesh()["cp"])
 
         if parallel_backend.tensor_parallel_enabled:
             # TODO(aryan): handle fp8 from TorchAO here
             model_specification.apply_tensor_parallel(
                 backend=parallel.ParallelBackendEnum.PTD,
-                device_mesh=parallel_backend.get_mesh("tp"),
+                device_mesh=parallel_backend.get_mesh()["tp"],
                 transformer=self.transformer,
             )
 
@@ -178,12 +172,9 @@ class SFTTrainer(Trainer):
             if self.args.parallel_backend == "accelerate":
                 raise NotImplementedError("Data sharding is not supported with Accelerate yet.")
 
-            if parallel_backend.data_replication_enabled:
-                logger.info("Applying HSDP to the model")
-            else:
-                logger.info("Applying FSDP to the model")
-
-            # Apply FSDP or HSDP
+            dp_method = "HSDP" if parallel_backend.data_replication_enabled else "FSDP"
+            logger.info(f"Applying {dp_method} to the model")
+            
             if parallel_backend.data_replication_enabled or parallel_backend.context_parallel_enabled:
                 dp_mesh_names = ("dp_replicate", "dp_shard_cp")
             else:
@@ -199,11 +190,9 @@ class SFTTrainer(Trainer):
                 device_mesh=parallel_backend.get_mesh()[dp_mesh_names],
             )
         elif parallel_backend.data_replication_enabled:
-            logger.info("Applying DDP to the model")
-
             if parallel_backend.get_mesh().ndim > 1:
                 raise ValueError("DDP not supported for > 1D parallelism")
-
+            logger.info("Applying DDP to the model")
             parallel_backend.apply_ddp(self.transformer, parallel_backend.get_mesh())
         else:
             parallel_backend.prepare_model(self.transformer)
@@ -749,41 +738,6 @@ class SFTTrainer(Trainer):
     def _evaluate(self) -> None:
         raise NotImplementedError("Evaluation has not been implemented yet.")
 
-    def _init_distributed(self) -> None:
-        world_size = int(os.environ.get("WORLD_SIZE", torch.cuda.device_count()))
-
-        # TODO(aryan): handle other backends
-        backend_cls: parallel.ParallelBackendType = parallel.get_parallel_backend_cls(self.args.parallel_backend)
-        self.state.parallel_backend = backend_cls(
-            world_size=world_size,
-            pp_degree=self.args.pp_degree,
-            dp_degree=self.args.dp_degree,
-            dp_shards=self.args.dp_shards,
-            cp_degree=self.args.cp_degree,
-            tp_degree=self.args.tp_degree,
-            backend="nccl",
-            timeout=self.args.init_timeout,
-            logging_dir=self.args.logging_dir,
-            output_dir=self.args.output_dir,
-            gradient_accumulation_steps=self.args.gradient_accumulation_steps,
-        )
-
-        if self.args.seed is not None:
-            self.state.parallel_backend.enable_determinism(self.args.seed)
-
-    def _init_logging(self) -> None:
-        logging._set_parallel_backend(self.state.parallel_backend)
-        logging.set_dependency_log_level(self.args.verbose, self.state.parallel_backend.is_local_main_process)
-        logger.info("Initialized FineTrainers")
-
-    def _init_trackers(self) -> None:
-        # TODO(aryan): handle multiple trackers
-        trackers = [self.args.report_to]
-        experiment_name = self.args.tracker_name or "finetrainers-experiment"
-        self.state.parallel_backend.initialize_trackers(
-            trackers, experiment_name=experiment_name, config=self._get_training_info(), log_dir=self.args.logging_dir
-        )
-
     def _init_directories_and_repositories(self) -> None:
         if self.state.parallel_backend.is_main_process:
             self.args.output_dir = Path(self.args.output_dir)
@@ -793,12 +747,6 @@ class SFTTrainer(Trainer):
             if self.args.push_to_hub:
                 repo_id = self.args.hub_model_id or Path(self.args.output_dir).name
                 self.state.repo_id = create_repo(token=self.args.hub_token, repo_id=repo_id, exist_ok=True).repo_id
-
-    def _init_config_options(self) -> None:
-        # Enable TF32 for faster training on Ampere GPUs: https://pytorch.org/docs/stable/notes/cuda.html#tensorfloat-32-tf32-on-ampere-devices
-        if self.args.allow_tf32 and torch.cuda.is_available():
-            torch.backends.cuda.matmul.allow_tf32 = True
-        torch.set_float32_matmul_precision(self.args.float32_matmul_precision)
 
     def _move_components_to_device(
         self, components: Optional[List[torch.nn.Module]] = None, device: Optional[Union[str, torch.device]] = None
